refactor: slice 적용하여 네이밍 수정
