+++ src/pages/ProductDetailPage.js
@@ -7,7 +7,7 @@ import Image from '../components/utils/Image';
 import Button from '../components/utils/Button';
 import PriceText from '../components/utils/PriceText';
 
-import { addItemToCart } from '../modules/cart';
+import { addItemToCart } from '../modules/cartSlice';
 import { addItemToCartRequest } from '../api/products';
 
 import styled, { css } from 'styled-components';
